Colors & Css
This module was insightful, covering the various ways to apply colors in CSS. From using hexadecimal (#hex) values and named colors to RGB, transparency, and more, it provided a comprehensive understanding of color implementation in web design. Learning about the nuances of each method enhances flexibility and creativity when styling websites.